src/main/java/com/qa/project/dto/ShopItemDTO.java
package com.qa.project.dto;

public class ShopItemDTO {
    private long id;
    private String itemName;
    private String itemCategory;
    private long itemQuantity;

    public ShopItemDTO() {

    }
    public ShopItemDTO(String itemName, String itemCategory, long itemQuantity) {
        this.itemName = itemName;
        this.itemCategory = itemCategory;
        this.itemQuantity = itemQuantity;

    }

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getItemName() {
        return itemName;
    }

    public void setItemName(String itemName) {
        this.itemName = itemName;
    }

    public String getItemCategory() {
        return itemCategory;
    }

    public void setItemCategory(String itemCategory) {
        this.itemCategory = itemCategory;
    }

    public long getItemQuantity() {
        return itemQuantity;
    }

    public void setItemQuantity(long itemQuantity) {
        this.itemQuantity = itemQuantity;
    }


}

